Major update
Trying to get this new one going. Im getting the standards and the base files set up.

After this, I need to continue work on the standards, then figure out how I want to send messages and stuff. Then I can focus on UI.

I really don't think the message thing will be that hard, it probably won't have notifcations or anything fancy with it, but it should work.