cloudFirestore.dart
Trying to get the cloudFirestore stuff set up.

I need to finish the last few methods for messages and for connecting to the other person. But after I do that I should have like 90% of the stuff done.

After that I will need to work on the UI for the two selection pages (status and ID) and then for the messages