login.dart
Trying to get everything set up to be consistent.

I think that after this one, then connector.dart shouldn't have that many to convert in it and home.dart shouldn't have that much in there yet.

I want to get login.dart at least setup tonight and maybe connector.dart. What would be ideal is to get all of them setup with the new stuff, that would be amazing.