home.dart and cloudFirestore.dart
cloudFirestore.dart: Created a possible thing for resetting the partner data. We will still need to create a button and make it to where we send them directly to the connector screen.

home.dart: Maybe this will work? If this does not work, then we can send the snapshots and make sure that something is coming through and make sure that everything is spelled the same.